Quarter 3/Academic Excellence Assignments/health_wellness_agent/health_wellness_agent_openai_sdk/pyproject.toml
[project]
name = "health-wellness"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.1.0",
    "pydantic>=2.11.7",
    "streamlit>=1.46.1",
]
